requirements.txt
asgiref==3.5.0
attrs==21.4.0
az-iranian-bank-gateways==1.6.15
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.12
django-crispy-forms==1.14.0
idna==3.3
isodate==0.6.1
lxml==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
pycryptodome==3.14.1
pytz==2021.3
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
zeep==4.1.0
aiohttp
asyncio
django-cleanup
persian-tools
django-cors-headers
djangorestframework
djangorestframework-simplejwt
encodings
psycopg2-binary

